enterprise_document_intelligence_rag/app/chunking/chunker.py
from langchain_text_splitters import RecursiveCharacterTextSplitter

def chunk_text(text: str) -> list[str]:
    if not isinstance(text, str):
        raise TypeError(f"Expected str, got {type(text)}")

    splitter = RecursiveCharacterTextSplitter(
        chunk_size=1000,
        chunk_overlap=150,
        separators=["\n\n", "\n", ". ", " ", ""]
    )

    return splitter.split_text(text)
