pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "app>=0.0.1",
    "datetime>=6.0",
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.2",
    "langchain-openai>=1.1.6",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20251107",
    "pdfplumber>=0.11.8",
    "pi-heif>=1.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "unstructured>=0.18.21",
    "unstructured-inference>=1.1.2",
    "uvicorn>=0.38.0",
]
